添加：权限管理, 当前用户仅有有权限的用户才可以对服务器进行关机、重启和开启终端。

+++ src/main/java/com/iecas/servermanageplatform/controller/ServerInfoController.java
@@ -59,7 +59,7 @@ public class ServerInfoController {
     @Auth
     @Logger("获取服务器详细信息")
     @GetMapping("/getServerInfo")
-    public CommonResult getServerInfo(@RequestParam Integer id){
+    public CommonResult getServerInfo(@RequestParam Long id){
         ServerInfo serverInfo = serverInfoService.getByIdEncryPwd(id);
         return new CommonResult().data(serverInfo).success();
     }
